Remove truffle dependencies (#75)
This PR removes the unneeded (for now) truffle dependency and configuration. I believe it was there from the subgraph template that contained its own contract that it needed to deploy. However, in our case, this isn't needed since the contract and deployment scripts are part of `dex-contracts`.

Eventually, we might need Truffle in order to setup integration tests for the subgraph, but in order to make it more obvious what is actually needed (for example, Infura configuration that was there will most likely not be needed) we first remove the unnecessary bits.

### Test Plan

Try the updated README instructions. This PR doesn't modify any of the graph code, just slims down the project structure.

### Commit History

* remove truffle dependencies

* update README

* remove truffle config

* fix fetch ABI script