Add xDAI deployment configuration (#91)
This PR adds the required configuration for xDAI deployment. Note that tokens don't index correctly because `Block and call handlers are not supported on the xdai network`. I created an issue #89 for adding tokens with the `TokenListing` event to work around this.

### Test Plan

Go to <https://thegraph.com/explorer/subgraph/gnosis/protocol-xdai> and try the following query:
```graphql
{
  solutions(first: 5) {
    batch {id}
    trades {
      sellVolume
      order{sellToken{address}}
    }
  }
}
```