start re-writing (#105)
This PR rewrites the Wasm to remove the AssemblyScript `start` function, and instead expose it as an exprot. This is done because the compiled AssemblyScript WebAssembly module calls exports in its `start` initialization method that **require** exports to work (for example `index/bigInt.pow` which requires the exported `memory`). This, unfortunately is not supported with the JavaScript WebAssembly API. To work around this, we disable the `start` function and add it as an export to be called manually **after** the module has been compiled.

### Test Plan

This PR can't be tested as is, as the Wasm start function requires exports that have currently not been implemented. It will be tested in a future PR.