Data driven type definitions (#102)
This PR continues the "data driven" type definition abstraction from the previous PR #101 (which was implemented for entity data). It is now also implemented for events, where they are specified by a "event definition" object. The advantage is that new events and entities can now be added by just adding new definitions to each of the entity and event definition objects in `definitions.ts`. Additionally, the event and entity type helpers are now submodules of the `runtime` module.

### Test Plan

The unit tests continue to pass!