Fix Travis (#127)
It seems like `docker-compose logs` is erroring with 

> Failed to execute script docker-compose

cf. https://travis-ci.com/github/gnosis/dex-subgraph/jobs/407146602

I was not able to find the root cause nor able reproduce it locally on my Mac but it seems reasonable to not fail travis if either `docker-compose logs` of `docker-compose stop` fails, therefore this PR is unsetting the `-e` flag.

Created #128 to track a longer term fix.