[Coverage 2] Annotate ASC source with trace instructions (#122)
This PR adds the code to annotate a given source file with trace instructions. We basically traverse the AST of the compiled assembly script and insert a `trace(filename, from_line, to_line)` statement in front of each meaningful impression.

Later we will instrument this invocation to keep track of which line in which file has been executed how many times.

### Test Plan

At the end of the PR stack we can run `yarn coverage` to generate a report.